refactor: コンポーネントのパスを変更
Template は ReviewPage でしか使わないため
