+++ src/components/review/TemplateList.tsx
@@ -1,7 +1,7 @@
 import { VStack } from '@chakra-ui/react';
 import React, { useState, useEffect } from 'react';
 
-import Template from '@/components/Template';
+import Template from '@/components/review/Template';
 
 const templates = [
   {
