+++ src/pages/review.tsx
@@ -6,9 +6,9 @@ import { BsFillChatDotsFill } from 'react-icons/bs';
 import { useParams, useHistory } from 'react-router';
 
 import Layout from '@/components/Layout';
-import TemplateList from '@/components/TemplateList';
 import DiffFiles from '@/components/review/DiffFiles';
 import ReviewTitle from '@/components/review/ReviewTitle';
+import TemplateList from '@/components/review/TemplateList';
 import useWidgets from '@/components/review/useWidgets';
 import { reviewer } from '@/data/dummyReviewer';
 import { useApi } from '@/hooks/useApi';
